+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## [0.1.0] 2026-07-13
+## [0.1.1] - 2026-07-14
+
+### Fixed
+- Flicker on substance page browse from Library
+
+### Changed
+- Modularization of some components
+
+
+## [0.1.0] - 2026-07-13
 
 ### Added
 - Changelog